backend/requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9      # required for file uploads in FastAPI

# ── ML / Inference ───────────────────────────────────────────────────────────
ultralytics>=8.3.0            # YOLOv8 — auto-downloads yolov8n.pt on first run
pillow>=10.0.0

# ── System metrics ───────────────────────────────────────────────────────────
psutil>=5.9.0
ortools>=9.10.0
httpx>=0.27.0
